Struct mo_gc::AppThread

pub struct AppThread;

An Application Thread, manages a thread-local reference to a tx channel

TODO: a version of spawn() is required that can be called from an existing mutator thread.

Methods

impl AppThread

fn spawn_from_gc<F, T>(tx: JournalSender, f: F) -> JoinHandle<T> where F: FnOnce() -> T, F: Send + 'static, T: Send + 'static

As thread::spawn but takes a journal Sender to initialize the thread_local instance with.