Struct mo_gc::GcAtomic

pub struct GcAtomic<T: Trace> {
    // some fields omitted
}

Atomic pointer type that points at a traceable object. This type is Sync and can be used to build concurrent data structures.

This type should be used inside data structures to reference other GC-managed objects, but provides interior mutability and atomic methods.

TODO: cas, swap etc for GcRoot and Gc

Methods

impl<T: Trace> GcAtomic<T>

fn null() -> GcAtomic<T>

Instantiate a new null pointer

fn new(value: T) -> GcAtomic<T>

Instantiate a new pointer, moving value to the heap. Writes to the journal.

fn load_into_root(&self, order: Ordering) -> GcRoot<T>

Root the pointer by loading it into a GcRoot<T>

Panics if order is Release or AcqRel.

fn load_into_gc(&self, order: Ordering) -> Gc<T>

Copy the pointer into a new Gc instance.

Panics if order is Release or AcqRel.

fn load_raw(&self, order: Ordering) -> *mut GcBox<T>

Fetch the current raw pointer value

Panics if order is Release or AcqRel.

fn store_from_root(&self, root: GcRoot<T>, order: Ordering)

Replace the current pointer value with the pointer from the given GcRoot.

Panics if order is Acquire or AcqRel.

fn store_from_gc(&self, gc: Gc<T>, order: Ordering)

Replace the current pointer value with the pointer from the given Gc.

Panics of order is Acquire or AcqRel.

fn store_raw(&self, ptr: *mut GcBox<T>, order: Ordering)

Replace the current pointer value with the given raw pointer

Panics if order is Acquire or AcqRel.