Struct mo_gc::GcRoot

pub struct GcRoot<T: Trace> {
    // some fields omitted
}

Root smart pointer, sends reference count changes to the journal.

Whenever a reference to an object on the heap must be retained on the stack, this type must be used. It's use will ensure that the object will be seen as a root.

Methods

impl<T: Trace> GcRoot<T>

fn new(value: T) -> GcRoot<T>

Put a new object on the heap and hand ownership to the GC, writing a reference count increment to the journal.