Struct mo_gc::GcThread

pub struct GcThread<S: StatsLogger> {
    // some fields omitted
}

The Garbage Collection thread handle.

Methods

impl GcThread<DefaultLogger>

fn spawn_gc() -> GcThread<DefaultLogger>

Spawn a GC thread with default parameters: a ParHeap and a DefaultLogger parallelized across all available CPUs.

impl<S: StatsLogger + 'static> GcThread<S>

fn spawn_gc_with<T>(num_threads: usize, mature: T, logger: S) -> GcThread<S> where T: CollectOps + Send + 'static

Run the GC on the current thread, spawning another thread to run the application function on. Returns the AppThread std::thread::Thread handle. Caller must provide a custom StatsLogger implementation and a CollectOps heap implementation.

fn spawn<F, T>(&self, f: F) -> JoinHandle<T> where F: FnOnce() -> T, F: Send + 'static, T: Send + 'static

Spawn an app thread that journals to the GC thread.

fn join(self) -> Result<S, Box<Any + Send + 'static>>

Wait for the GC thread to finish. On success, returns the object that implements StatsLogger for the calling thread to examine.