Struct mo_gc::ParHeap

pub struct ParHeap {
    // some fields omitted
}

This references all known GC-managed objects and handles marking and sweeping; parallel mark and sweep version.

Methods

impl ParHeap

fn new(num_threads: usize) -> ParHeap

In this heap implementation, work is split out into a thread pool. There is no knowing, though, how much work each split actually represents. One thread may receive a disproportionate amount of tracing or sweeping.