Struct mo_gc::Receiver

pub struct Receiver<T> {
    // some fields omitted
}

A journal reader type which can be sent to another thread

Methods

impl<T> Receiver<T>

fn try_recv(&self) -> Result<T, RecvResult>

Read a value from the queue if there is one available, otherwise return without blocking

fn iter_until_empty(&mut self) -> EmptyIter<T>

Make an Iterator that returns values until the queue is empty or disconnected.

fn is_disconnected(&self) -> bool

Has the Sender hung up?