Struct mo_gc::TraceStack

pub struct TraceStack {
    // some fields omitted
}

A type that contains a stack of objects to trace into. This type is separated out from the main Heap type so that different collection strategies can be implemented without affecting the client code. The Trace trait depends only this type, then, and not the whole Heap type.

Methods

impl TraceStack

fn new() -> TraceStack

fn push(&mut self, obj: Object)

fn pop(&mut self) -> Option<Object>

fn from_roots(&mut self, slice: &[Object])