Struct mo_gc::YoungHeap

pub struct YoungHeap<S: StatsLogger, T: CollectOps + Send> {
    // some fields omitted
}

Type that composes all the things we need to run garbage collection on young generation objects.

The roots trie maps object addresses to their reference counts, vtables and NEW object flags.

During tracing, positive reference count objects and non-NEW objects are considered possible roots and only NEW objects are considered for marking and sweeping. Entries can be both roots and NEW.

Collection is run in a thread pool across all CPUs by default by sharding the root trie across threads.

Methods

impl<S: StatsLogger, T: CollectOps + Send> YoungHeap<S, T>

fn new(num_threads: usize, mature: T, logger: S) -> YoungHeap<S, T>

Create a new young generation heap and roots reference count tracker

fn add_journal(&mut self, recv: EntryReceiver)

Add a new journal to the list of journals to read

fn num_journals(&self) -> usize

Returns the number of journals currently connected to the GC

fn read_journals(&mut self) -> usize

Read all journals for a number of iterations, updating the roots and keeping a reference count increment for each, and putting decrements into the deferred buffer.

This function is single-threaded and is the biggest GC throughput bottleneck. Setting a value in the trie is slow compared to allocation and writing/reading the journal.

Easily consumes 80% of linear GC time. TODO: parallelize this function.

Returns the number of journal entries read.

fn minor_collection(&mut self, pool: &mut Pool) -> usize

Do a young generation collection. Returns the number of new objects in the young generation heap.

fn major_collection(&mut self, pool: &mut Pool)

Do a major collection, moving NEW objects to the mature heap and tracing the mature heap

fn logger(&mut self) -> &mut S

Return a reference to the logger

fn shutdown(self) -> S

Call to return the logger on shutdown