Trait mo_gc::Trace

pub unsafe trait Trace {
    fn traversible(&self) -> bool { ... }
    unsafe fn trace(&self, _stack: &mut TraceStack) { ... }
}

Trace trait. Every type that can be managed by the GC must implement this trait. This trait is unsafe in that incorrectly implementing it can cause Undefined Behavior.

Provided Methods

fn traversible(&self) -> bool

If the type can contain GC managed pointers, this must return true

unsafe fn trace(&self, _stack: &mut TraceStack)

If the type can contain GC managed pointers, this must visit each pointer.

This function must be thread-safe!

It must read a snapshot of the data structure it is implemented for.